@@ -1,29 +1,44 @@
 {
   "name": "fast-cashier",
   "version": "0.0.0",
+  "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview --port 5050",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "test:unit": "vitest",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "pinia": "^2.0.11",
-    "primeflex": "^3.1.3",
-    "primeicons": "^5.0.0",
-    "primevue": "^3.12.1",
-    "vue": "^3.2.31",
-    "vue-router": "^4.0.12"
+    "@primevue/themes": "^4.0.7",
+    "pinia": "^2.2.2",
+    "primevue": "^4.0.7",
+    "vue": "^3.5.6",
+    "vue-router": "^4.4.5"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.1.0",
-    "@vitejs/plugin-vue": "^2.2.2",
-    "@vue/eslint-config-prettier": "^7.0.0",
-    "eslint": "^8.5.0",
-    "eslint-plugin-vue": "^8.2.0",
-    "prettier": "^2.5.1",
-    "sass": "^1.49.9",
-    "sass-loader": "^12.6.0",
-    "vite": "^2.8.4"
+    "@rushstack/eslint-patch": "^1.10.4",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/jsdom": "^21.1.7",
+    "@types/node": "^22.5.5",
+    "@vitejs/plugin-vue": "^5.1.3",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/test-utils": "^2.4.6",
+    "@vue/tsconfig": "^0.5.1",
+    "eslint": "^8.57.1",
+    "eslint-plugin-vue": "^9.28.0",
+    "jsdom": "^25.0.0",
+    "npm-run-all2": "^6.2.3",
+    "prettier": "^3.3.3",
+    "sass": "^1.78.0",
+    "typescript": "~5.6.2",
+    "vite": "^5.4.6",
+    "vitest": "^2.1.1",
+    "vue-tsc": "^2.1.6"
   }
 }
