@@ -16,28 +16,26 @@
   </main>
 </template>
 
-<script>
-import CategoriesList from "../components/categories/CategoriesList.vue";
-import MainKeyboard from "../components/keyboard/MainKeyboard.vue";
-import RecordList from "../components/record/RecordList.vue";
-import KeyboardDisplay from "../components/keyboard/KeyboardDisplay.vue";
-import { keyboardStore } from "../stores/keyboard";
-import { mapActions } from "pinia";
-export default {
-  components: { CategoriesList, MainKeyboard, RecordList, KeyboardDisplay },
-  mounted() {
-    window.addEventListener("keyup", this.dispatchKeyEvent);
-  },
-  beforeUnmount() {
-    window.removeEventListener("keyup", this.dispatchKeyEvent);
-  },
-  methods: {
-    ...mapActions(keyboardStore, ["emitKey", "hasKey"]),
-    dispatchKeyEvent: function (event) {
-      this.emitKey(event.key);
-    },
-  },
-};
+<script setup lang="ts">
+import CategoriesList from '@/components/categories/CategoriesList.vue'
+import MainKeyboard from '@/components/keyboard/MainKeyboard.vue'
+import RecordList from '@/components/record/RecordList.vue'
+import KeyboardDisplay from '@/components/keyboard/KeyboardDisplay.vue'
+import { useKeyboardStore } from '@/stores/keyboard'
+import { onMounted, onBeforeUnmount } from 'vue'
+
+const keyboardStore = useKeyboardStore()
+
+const dispatchKeyEvent = (event: KeyboardEvent) => {
+  keyboardStore.emitKey(event.key)
+}
+
+onMounted(() => {
+  window.addEventListener('keyup', dispatchKeyEvent)
+})
+onBeforeUnmount(() => {
+  window.removeEventListener('keyup', dispatchKeyEvent)
+})
 </script>
 
 <style lang="scss">
@@ -47,7 +45,7 @@ export default {
   display: flex;
   flex-direction: column;
 
-  >div {
+  > div {
     padding: 10px;
     border: solid red 1px;
   }
@@ -68,7 +66,7 @@ export default {
       grid-column-start: 3;
     }
 
-    >div {
+    > div {
       padding: 10px;
       border: solid red 1px;
     }
