@@ -1,11 +1,11 @@
-<script setup>
-import { RouterView } from "vue-router";
-import AppTopBar from "./components/AppTopBar.vue";
-import AppFooter from "./components/AppFooter.vue";
+<script setup lang="ts">
+import { RouterView } from 'vue-router'
+import AppTopBar from '@/components/AppTopBar.vue'
+import AppFooter from '@/components/AppFooter.vue'
 </script>
 
 <template>
-  <div>
+  <div id="fc-app-container">
     <AppTopBar />
     <div class="layout-main-container">
       <div class="layout-main">
@@ -15,9 +15,3 @@ import AppFooter from "./components/AppFooter.vue";
     </div>
   </div>
 </template>
-
-<style lang="scss">
-@import "primevue/resources/themes/md-dark-deeppurple/theme.css";
-@import "primevue/resources/primevue.min.css";
-@import "primeicons/primeicons.css";
-</style>
