@@ -1,24 +1,14 @@
 <template>
   <div class="categories-list">
-    <CategoryItem
-      v-for="category in categories"
-      :key="category.id"
-      :category="category"
-    />
+    <CategoryItem v-for="category in DEFAULT_CATEGORIES" :key="category.id" :category="category" />
   </div>
 </template>
-<script>
-import { DEFAULT_CATEGORIES } from "../../config";
-import CategoryItem from "./CategoryItem.vue";
-export default {
-  components: { CategoryItem },
-  data: function () {
-    return {
-      categories: DEFAULT_CATEGORIES,
-    };
-  },
-};
+
+<script setup lang="ts">
+import { DEFAULT_CATEGORIES } from '@/domain/config'
+import CategoryItem from '@/components/categories/CategoryItem.vue'
 </script>
+
 <style lang="scss" scoped>
 .categories-list {
   display: grid;
