@@ -1,124 +1,101 @@
-<script>
-import { mapActions } from "pinia";
-import { ALLOWED_ACTIONS, KEY_TYPES } from "../../config";
-import { keyboardStore } from "../../stores/keyboard";
-import KeyboardButton from "./KeyboardButton.vue";
-export default {
-  components: { KeyboardButton },
-  data() {
-    return {
-      leftValue: 0,
-      operator: "",
-      rightValue: null,
-      result: 0,
-      hasResult: false,
-      actions: Object.freeze([
-        {
-          name: "key-clear",
-          value: ALLOWED_ACTIONS.DELETE,
-          display: "C",
-          type: KEY_TYPES.ACTION,
-        },
-        {
-          name: "key-return",
-          value: ALLOWED_ACTIONS.BACKSPACE,
-          display: "<-",
-          type: KEY_TYPES.ACTION,
-        },
-        {
-          name: "key-equals",
-          value: ALLOWED_ACTIONS.ENTER,
-          display: "=",
-          type: KEY_TYPES.ACTION,
-        },
-      ]),
-      keys: Object.freeze([
-        { name: "key-1", value: 1, display: "1", type: KEY_TYPES.DIGIT },
-        { name: "key-2", value: 2, display: "2", type: KEY_TYPES.DIGIT },
-        { name: "key-3", value: 3, display: "3", type: KEY_TYPES.DIGIT },
-        { name: "key-4", value: 4, display: "4", type: KEY_TYPES.DIGIT },
-        { name: "key-5", value: 5, display: "5", type: KEY_TYPES.DIGIT },
-        { name: "key-6", value: 6, display: "6", type: KEY_TYPES.DIGIT },
-        { name: "key-7", value: 7, display: "7", type: KEY_TYPES.DIGIT },
-        { name: "key-8", value: 8, display: "8", type: KEY_TYPES.DIGIT },
-        { name: "key-9", value: 9, display: "9", type: KEY_TYPES.DIGIT },
-        { name: "key-0", value: 0, display: "0", type: KEY_TYPES.DIGIT },
-        {
-          name: "key-00",
-          value: ALLOWED_ACTIONS.HUNDRED,
-          display: ALLOWED_ACTIONS.HUNDRED,
-          type: KEY_TYPES.ACTION,
-        },
-        {
-          name: "key-000",
-          value: ALLOWED_ACTIONS.THOUSAND,
-          display: ALLOWED_ACTIONS.THOUSAND,
-          type: KEY_TYPES.ACTION,
-        },
-        {
-          name: "key-multiply",
-          value: "*",
-          display: "*",
-          type: KEY_TYPES.OPERATION,
-        },
-        {
-          name: "key-plus",
-          value: "+",
-          display: "+",
-          type: KEY_TYPES.OPERATION,
-        },
-      ]),
-    };
-  },
-  mounted() {
-    this.actions.forEach((key) => this.registerKey(key.value, key));
-    this.keys.forEach((key) => this.registerKey(key.value, key));
-  },
-  beforeUnmount() {
-    this.keys.forEach((key) => this.unRegisterKey(key.value, key));
-    this.actions.forEach((key) => this.unRegisterKey(key.value, key));
-  },
-  methods: {
-    ...mapActions(keyboardStore, [
-      "keyPressed",
-      "registerKey",
-      "unRegisterKey",
-    ]),
-  },
-  computed: {
-    displayValue() {
-      const r = this.rightValue > 0 ? this.rightValue : "";
-      if (this.operator === "") return `${this.leftValue}`;
-      else return `${this.leftValue} ${this.operator} ${r}`;
-    },
-  },
-};
-</script>
-
 <template>
   <div class="keyboard">
     <div class="actions">
       <KeyboardButton
-        v-for="key in this.actions"
+        v-for="key in actions"
         :key="key.name"
         :class="`p-button-outlined p-button-secondary ${key.name}`"
         :label="key.display"
         :button="key"
-        @click="keyPressed(key)"
+        @click="keyboardStore.keyPressed(key)"
       />
     </div>
     <div class="numbers">
       <KeyboardButton
-        v-for="key in this.keys"
+        v-for="key in keys"
         :key="key.name"
         :class="`p-button-outlined p-button-secondary ${key.name}`"
         :label="key.display"
         :button="key"
-        @click="keyPressed(key)"
+        @click="keyboardStore.keyPressed(key)"
       />
     </div>
   </div>
 </template>
+
+<script setup lang="ts">
+import { onBeforeMount, onMounted } from 'vue'
+import { ALLOWED_ACTIONS, KEY_TYPES } from '@/domain/config'
+import { useKeyboardStore } from '@/stores/keyboard'
+import KeyboardButton from '@/components/keyboard/KeyboardButton.vue'
+import type { KeyMetadata } from '@/domain/key-metadata'
+
+const keyboardStore = useKeyboardStore()
+
+const actions: Array<KeyMetadata> = [
+  {
+    name: 'key-clear',
+    value: ALLOWED_ACTIONS.DELETE,
+    display: 'C',
+    type: KEY_TYPES.ACTION
+  },
+  {
+    name: 'key-return',
+    value: ALLOWED_ACTIONS.BACKSPACE,
+    display: '<-',
+    type: KEY_TYPES.ACTION
+  },
+  {
+    name: 'key-equals',
+    value: ALLOWED_ACTIONS.ENTER,
+    display: '=',
+    type: KEY_TYPES.ACTION
+  }
+]
+const keys: Array<KeyMetadata> = [
+  { name: 'key-1', value: 1, display: '1', type: KEY_TYPES.DIGIT },
+  { name: 'key-2', value: 2, display: '2', type: KEY_TYPES.DIGIT },
+  { name: 'key-3', value: 3, display: '3', type: KEY_TYPES.DIGIT },
+  { name: 'key-4', value: 4, display: '4', type: KEY_TYPES.DIGIT },
+  { name: 'key-5', value: 5, display: '5', type: KEY_TYPES.DIGIT },
+  { name: 'key-6', value: 6, display: '6', type: KEY_TYPES.DIGIT },
+  { name: 'key-7', value: 7, display: '7', type: KEY_TYPES.DIGIT },
+  { name: 'key-8', value: 8, display: '8', type: KEY_TYPES.DIGIT },
+  { name: 'key-9', value: 9, display: '9', type: KEY_TYPES.DIGIT },
+  { name: 'key-0', value: 0, display: '0', type: KEY_TYPES.DIGIT },
+  {
+    name: 'key-00',
+    value: ALLOWED_ACTIONS.HUNDRED,
+    display: ALLOWED_ACTIONS.HUNDRED,
+    type: KEY_TYPES.ACTION
+  },
+  {
+    name: 'key-000',
+    value: ALLOWED_ACTIONS.THOUSAND,
+    display: ALLOWED_ACTIONS.THOUSAND,
+    type: KEY_TYPES.ACTION
+  },
+  {
+    name: 'key-multiply',
+    value: '*',
+    display: '*',
+    type: KEY_TYPES.OPERATION
+  },
+  {
+    name: 'key-plus',
+    value: '+',
+    display: '+',
+    type: KEY_TYPES.OPERATION
+  }
+]
+
+onMounted(() => {
+  ;[...actions, ...keys].forEach((key) => keyboardStore.registerKey(key.value, key))
+})
+onBeforeMount(() => {
+  ;[...actions, ...keys].forEach((key) => keyboardStore.unRegisterKey(key.value))
+})
+</script>
 
 <style scoped lang="scss">
 .display {
